Renamed main to MainForm to keep form naming consistent.
