+++ Program.cs
@@ -17,7 +17,7 @@ namespace NextGraphics
 		{
 			Application.EnableVisualStyles();
 			Application.SetCompatibleTextRenderingDefault(false);
-			Application.Run(new Main());
+			Application.Run(new MainForm());
 		}	
 		
 	}
